Flatten e_particles.h
